move routing.rs and all dependent types back into near_network
