+++ chain/network-primitives/Cargo.toml
@@ -7,14 +7,11 @@ edition = "2018"
 [dependencies]
 actix = "=0.11.0-beta.2"
 tokio = { version = "1.1", features = ["full"] }
-futures = "0.3"
 chrono = { version = "0.4.4", features = ["serde"] }
 borsh = "0.8.1"
 serde = { version = "1", features = [ "derive" ] }
 strum = { version = "0.20", features = ["derive"] }
 tracing = "0.1.13"
-conqueue = "0.4.0"
-byteorder = "1.2"
 
 near-crypto = { path = "../../core/crypto" }
 near-primitives = { path = "../../core/primitives" }
