+++ chain/network-primitives/src/lib.rs
@@ -1,2 +1,1 @@
-pub mod routing;
 pub mod types;
