+++ chain/network/src/routing.rs
@@ -3,13 +3,15 @@ use std::ops::Sub;
 use std::sync::{Arc, Mutex};
 use std::time::Instant;
 
+use borsh::{BorshDeserialize, BorshSerialize};
+use byteorder::{LittleEndian, WriteBytesExt};
 use cached::{Cached, SizedCache};
 use chrono;
 use tracing::{trace, warn};
 
-use crate::cache::RouteBackCache;
+use near_crypto::{SecretKey, Signature};
 use near_metrics;
-use near_primitives::hash::CryptoHash;
+use near_primitives::hash::{hash, CryptoHash};
 use near_primitives::network::{AnnounceAccount, PeerId};
 use near_primitives::types::AccountId;
 use near_primitives::utils::index_to_bytes;
@@ -20,14 +22,13 @@ use near_store::{
 
 use crate::metrics;
 use crate::{
+    cache::RouteBackCache,
     types::{PeerIdOrHash, Ping, Pong},
     utils::cache_to_hashmap,
 };
 use conqueue::{QueueReceiver, QueueSender};
 #[cfg(feature = "delay_detector")]
 use delay_detector::DelayDetector;
-
-pub use near_network_primitives::routing::*;
 
 const ANNOUNCE_ACCOUNT_CACHE_SIZE: usize = 10_000;
 const ROUTE_BACK_CACHE_SIZE: u64 = 100_000;
@@ -36,6 +37,222 @@ const ROUTE_BACK_CACHE_REMOVE_BATCH: u64 = 100;
 const PING_PONG_CACHE_SIZE: usize = 1_000;
 const ROUND_ROBIN_MAX_NONCE_DIFFERENCE_ALLOWED: usize = 10;
 const ROUND_ROBIN_NONCE_CACHE_SIZE: usize = 10_000;
+/// Routing table will clean edges if there is at least one node that is not reachable
+/// since `SAVE_PEERS_MAX_TIME` seconds. All peers disconnected since `SAVE_PEERS_AFTER_TIME`
+/// seconds will be removed from cache and persisted in disk.
+pub const SAVE_PEERS_MAX_TIME: u64 = 7_200;
+pub const SAVE_PEERS_AFTER_TIME: u64 = 3_600;
+/// Graph implementation supports up to 128 peers.
+pub const MAX_NUM_PEERS: usize = 128;
+
+/// Information that will be ultimately used to create a new edge.
+/// It contains nonce proposed for the edge with signature from peer.
+#[derive(Clone, BorshSerialize, BorshDeserialize, PartialEq, Eq, Debug, Default)]
+pub struct EdgeInfo {
+    pub nonce: u64,
+    pub signature: Signature,
+}
+
+impl EdgeInfo {
+    pub fn new(peer0: PeerId, peer1: PeerId, nonce: u64, secret_key: &SecretKey) -> Self {
+        let (peer0, peer1) = Edge::key(peer0, peer1);
+        let data = Edge::build_hash(&peer0, &peer1, nonce);
+        let signature = secret_key.sign(data.as_ref());
+        Self { nonce, signature }
+    }
+}
+
+/// Status of the edge
+#[derive(BorshSerialize, BorshDeserialize, Clone, PartialEq, Eq, Debug, Hash)]
+pub enum EdgeType {
+    Added,
+    Removed,
+}
+
+/// Edge object. Contains information relative to a new edge that is being added or removed
+/// from the network. This is the information that is required.
+#[derive(BorshSerialize, BorshDeserialize, Clone, Debug, PartialEq, Eq)]
+pub struct Edge {
+    /// Since edges are not directed `peer0 < peer1` should hold.
+    pub peer0: PeerId,
+    pub peer1: PeerId,
+    /// Nonce to keep tracking of the last update on this edge.
+    /// It must be even
+    pub nonce: u64,
+    /// Signature from parties validating the edge. These are signature of the added edge.
+    signature0: Signature,
+    signature1: Signature,
+    /// Info necessary to declare an edge as removed.
+    /// The bool says which party is removing the edge: false for Peer0, true for Peer1
+    /// The signature from the party removing the edge.
+    removal_info: Option<(bool, Signature)>,
+}
+
+impl Edge {
+    /// Create an addition edge.
+    pub fn new(
+        peer0: PeerId,
+        peer1: PeerId,
+        nonce: u64,
+        signature0: Signature,
+        signature1: Signature,
+    ) -> Self {
+        let (peer0, signature0, peer1, signature1) = if peer0 < peer1 {
+            (peer0, signature0, peer1, signature1)
+        } else {
+            (peer1, signature1, peer0, signature0)
+        };
+
+        Self { peer0, peer1, nonce, signature0, signature1, removal_info: None }
+    }
+
+    /// Build a new edge with given information from the other party.
+    pub fn build_with_secret_key(
+        peer0: PeerId,
+        peer1: PeerId,
+        nonce: u64,
+        secret_key: &SecretKey,
+        signature1: Signature,
+    ) -> Self {
+        let hash = if peer0 < peer1 {
+            Edge::build_hash(&peer0, &peer1, nonce)
+        } else {
+            Edge::build_hash(&peer1, &peer0, nonce)
+        };
+        let signature0 = secret_key.sign(hash.as_ref());
+        Edge::new(peer0, peer1, nonce, signature0, signature1)
+    }
+
+    /// Create the remove edge change from an added edge change.
+    pub fn remove_edge(&self, me: PeerId, sk: &SecretKey) -> Self {
+        assert_eq!(self.edge_type(), EdgeType::Added);
+        let mut edge = self.clone();
+        edge.nonce += 1;
+        let me = edge.peer0 == me;
+        let hash = edge.hash();
+        let signature = sk.sign(hash.as_ref());
+        edge.removal_info = Some((me, signature));
+        edge
+    }
+
+    /// Build the hash of the edge given its content.
+    /// It is important that peer0 < peer1 at this point.
+    fn build_hash(peer0: &PeerId, peer1: &PeerId, nonce: u64) -> CryptoHash {
+        let mut buffer = Vec::<u8>::new();
+        let peer0: Vec<u8> = peer0.clone().into();
+        buffer.extend_from_slice(peer0.as_slice());
+        let peer1: Vec<u8> = peer1.clone().into();
+        buffer.extend_from_slice(peer1.as_slice());
+        buffer.write_u64::<LittleEndian>(nonce).unwrap();
+        hash(buffer.as_slice())
+    }
+
+    fn hash(&self) -> CryptoHash {
+        Edge::build_hash(&self.peer0, &self.peer1, self.nonce)
+    }
+
+    fn prev_hash(&self) -> CryptoHash {
+        Edge::build_hash(&self.peer0, &self.peer1, self.nonce - 1)
+    }
+
+    pub fn verify(&self) -> bool {
+        if self.peer0 > self.peer1 {
+            return false;
+        }
+
+        match self.edge_type() {
+            EdgeType::Added => {
+                let data = self.hash();
+
+                self.removal_info.is_none()
+                    && self.signature0.verify(data.as_ref(), &self.peer0.public_key())
+                    && self.signature1.verify(data.as_ref(), &self.peer1.public_key())
+            }
+            EdgeType::Removed => {
+                // nonce should be an even positive number
+                if self.nonce == 0 {
+                    return false;
+                }
+
+                // Check referring added edge is valid.
+                let add_hash = self.prev_hash();
+                if !self.signature0.verify(add_hash.as_ref(), &self.peer0.public_key())
+                    || !self.signature1.verify(add_hash.as_ref(), &self.peer1.public_key())
+                {
+                    return false;
+                }
+
+                if let Some((party, signature)) = &self.removal_info {
+                    let peer = if *party { &self.peer0 } else { &self.peer1 };
+                    let del_hash = self.hash();
+                    signature.verify(del_hash.as_ref(), &peer.public_key())
+                } else {
+                    false
+                }
+            }
+        }
+    }
+
+    pub fn key(peer0: PeerId, peer1: PeerId) -> (PeerId, PeerId) {
+        if peer0 < peer1 {
+            (peer0, peer1)
+        } else {
+            (peer1, peer0)
+        }
+    }
+
+    /// Helper function when adding a new edge and we receive information from new potential peer
+    /// to verify the signature.
+    pub fn partial_verify(peer0: PeerId, peer1: PeerId, edge_info: &EdgeInfo) -> bool {
+        let pk = peer1.public_key();
+        let (peer0, peer1) = Edge::key(peer0, peer1);
+        let data = Edge::build_hash(&peer0, &peer1, edge_info.nonce);
+        edge_info.signature.verify(data.as_ref(), &pk)
+    }
+
+    fn get_pair(&self) -> (PeerId, PeerId) {
+        (self.peer0.clone(), self.peer1.clone())
+    }
+
+    /// It will be considered as a new edge if the nonce is odd, otherwise it is canceling the
+    /// previous edge.
+    pub fn edge_type(&self) -> EdgeType {
+        if self.nonce % 2 == 1 {
+            EdgeType::Added
+        } else {
+            EdgeType::Removed
+        }
+    }
+
+    /// Next nonce of valid addition edge.
+    pub fn next_nonce(nonce: u64) -> u64 {
+        if nonce % 2 == 1 {
+            nonce + 2
+        } else {
+            nonce + 1
+        }
+    }
+
+    /// Next nonce of valid addition edge.
+    pub fn next(&self) -> u64 {
+        Edge::next_nonce(self.nonce)
+    }
+
+    pub fn contains_peer(&self, peer_id: &PeerId) -> bool {
+        self.peer0 == *peer_id || self.peer1 == *peer_id
+    }
+
+    /// Find a peer id in this edge different from `me`.
+    pub fn other(&self, me: &PeerId) -> Option<PeerId> {
+        if self.peer0 == *me {
+            Some(self.peer1.clone())
+        } else if self.peer1 == *me {
+            Some(self.peer0.clone())
+        } else {
+            None
+        }
+    }
+}
 
 pub struct EdgeVerifierHelper {
     /// Shared version of edges_info used by multiple threads
@@ -278,7 +495,7 @@ impl RoutingTable {
     }
 
     fn add_edge(&mut self, edge: Edge) -> bool {
-        let key = (edge.peer0.clone(), edge.peer1.clone());
+        let key = edge.get_pair();
 
         if self.find_nonce(&key) >= edge.nonce {
             // We already have a newer information about this edge. Discard this information.
@@ -305,7 +522,7 @@ impl RoutingTable {
         let total = edges.len();
 
         for edge in edges {
-            let key = (&edge.peer0, &edge.peer1);
+            let key = edge.get_pair();
 
             self.touch(&key.0);
             self.touch(&key.1);
@@ -521,6 +738,12 @@ impl RoutingTable {
 
 pub struct ProcessEdgeResult {
     pub new_edge: bool,
+}
+
+#[derive(Debug)]
+pub struct RoutingTableInfo {
+    pub account_peers: HashMap<AccountId, PeerId>,
+    pub peer_forwarding: HashMap<PeerId, Vec<PeerId>>,
 }
 
 #[derive(Clone)]
